Update
- Added favicon
- Added follows
- Added favorites
- Added CSRF tokens
- Added error page
- Added 'all uploads' page
- Added follows page
- Added favorites page
- Added users search function
- Added view counts
- Added default profile picture

+++ edit.php
@@ -25,42 +25,46 @@ $page_title = "Upload - Edit Post: " . $upload_data->upload_title;
 $img_exts = array('jpg', 'png', 'PNG', 'gif');
 
 if(Input::exists($_POST, 'edit')) {
-  $validate = new Validate();
+  if(Hash::checkToken(Input::get('token'), 'token')) {
+    $validate = new Validate();
 
-  $validation = $validate->check($_POST, array(
-    'upload_title' => array(
-      'required' => true,
-      'min' => 1,
-      'max' => 150
-    ),
-    'upload_description' => array(
-      'max' => 500
-    )
-  ));
-
-  if($validation->passed()) {
-    if($upload->editUpload($id, array(
-      'upload_title' => escape(Input::get('upload_title')),
-      'upload_description' => escape(Input::get('upload_description'))
-    ))) {
-      Redirect::to(BASE_URL);
-    } else {
-      $validation->addError("Unable to edit upload");
+    $validation = $validate->check($_POST, array(
+      'upload_title' => array(
+        'required' => true,
+        'min' => 1,
+        'max' => 150
+      ),
+      'upload_description' => array(
+        'max' => 500
+      )
+    ));
+  
+    if($validation->passed()) {
+      if($upload->editUpload($id, array(
+        'upload_title' => escape(Input::get('upload_title')),
+        'upload_description' => escape(Input::get('upload_description'))
+      ))) {
+        Redirect::to(BASE_URL);
+      } else {
+        $validation->addError("Unable to edit upload");
+      }
     }
   }
 }
 
 if(Input::exists($_POST, 'delete_upload')) {
-  $upload_dir = "uploads/uploads/";
-  $upload_name = $upload_data->upload_file;
-  $file_to_delete = $upload_dir . $upload_name;
-
-  if($upload->deleteUpload($id)) {
-    if(file_exists($file_to_delete)) {
-      unlink($file_to_delete);
+  if(Hash::checkToken(Input::get('delete-token'), 'delete-token')) {
+    $upload_dir = "uploads/uploads/";
+    $upload_name = $upload_data->upload_file;
+    $file_to_delete = $upload_dir . $upload_name;
+  
+    if($upload->deleteUpload($id)) {
+      if(file_exists($file_to_delete)) {
+        unlink($file_to_delete);
+      }
+      
+      Redirect::to(BASE_URL);
     }
-    
-    Redirect::to(BASE_URL);
   }
 }
